refactor structure, add final design
