+++ Laboratory1/app/src/main/java/com/example/laboratory1/FifthActivity.kt
@@ -1,14 +1,12 @@
-package com.example.laboratory1.presenter
+package com.example.laboratory1
 
 import android.annotation.SuppressLint
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.laboratory1.R
 import com.example.laboratory1.databinding.ActivityFifthBinding
 
-import com.example.laboratory1.presenter.game.SecondActivity
 import com.example.laboratory1.recycleView.RecycleAdapter
 import com.example.laboratory1.utils.AppConstants.bottomNavigation
 import com.example.laboratory1.utils.AppFunctions.Companion.getCurrentDateTime
