+++ Laboratory1/app/src/main/java/com/example/laboratory1/FourthActivity.kt
@@ -1,11 +1,9 @@
-package com.example.laboratory1.presenter
+package com.example.laboratory1
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.example.laboratory1.R
 import com.example.laboratory1.databinding.ActivityFourthBinding
-import com.example.laboratory1.presenter.game.SecondActivity
 import com.example.laboratory1.utils.AppConstants.bottomNavigation
 import com.example.laboratory1.utils.AppFunctions.Companion.getPosts
