+++ Laboratory1/app/src/main/java/com/example/laboratory1/SecondActivity.kt
@@ -1,14 +1,10 @@
-package com.example.laboratory1.presenter.game
+package com.example.laboratory1
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.example.laboratory1.R
 import com.example.laboratory1.databinding.ActivitySecondBinding
-import com.example.laboratory1.presenter.FourthActivity
-import com.example.laboratory1.presenter.ThirdActivity
-import com.example.laboratory1.presenter.FifthActivity
-import com.example.laboratory1.presenter.SixthActivity
+import com.example.laboratory1.game.GameActivity
 import com.example.laboratory1.utils.AppConstants.bottomNavigation
 
 open class SecondActivity : AppCompatActivity() {
