+++ Laboratory1/app/src/main/java/com/example/laboratory1/ThirdActivity.kt
@@ -1,13 +1,11 @@
-package com.example.laboratory1.presenter
+package com.example.laboratory1
 
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
-import com.example.laboratory1.R
 import com.example.laboratory1.databinding.ActivityThirdBinding
-import com.example.laboratory1.presenter.game.SecondActivity
 import com.example.laboratory1.utils.AppConstants.bottomNavigation
