+++ Laboratory1/app/src/main/java/com/example/laboratory1/game/GameManager.kt
@@ -1,6 +1,5 @@
-package com.example.laboratory1.presenter.game
+package com.example.laboratory1.game
 
-import com.example.laboratory1.presenter.game.model.Position
 import com.example.laboratory1.utils.WinningLine
 
 class GameManager {
