+++ Laboratory1/app/src/main/res/layout/activity_fourth.xml
@@ -5,14 +5,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/pink_100"
-    tools:context=".presenter.FourthActivity">
-
-    <ImageView
-        android:id="@+id/imageView2"
-        android:layout_width="match_parent"
-        android:layout_height="674dp"
-        app:srcCompat="@drawable/background_01" />
+    android:background="@drawable/background_01"
+    tools:context=".FourthActivity">
 
     <ImageView
         android:id="@+id/imageView3"
