+++ Laboratory1/app/src/main/res/layout/activity_game.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".presenter.game.GameActivity">
+    tools:context=".game.GameActivity">
 
     <TextView
         android:id="@+id/player_one_score"
