+++ Laboratory1/app/src/main/res/layout/activity_second.xml
@@ -1,21 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:app="http://schemas.android.com/apk/res-auto"
-xmlns:tools="http://schemas.android.com/tools"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-tools:context=".presenter.game.SecondActivity">
-
-    <ImageView
-        android:id="@+id/imageView6"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:srcCompat="@drawable/background_02" />
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@drawable/background_02"
+    tools:context=".SecondActivity">
 
     <ImageView
         android:id="@+id/imageView7"
-        android:foregroundGravity="center"
         android:layout_width="325dp"
         android:layout_height="237dp"
         android:layout_above="@+id/bottom_navigation"
@@ -30,6 +23,7 @@ tools:context=".presenter.game.SecondActivity">
         android:layout_marginEnd="33dp"
         android:layout_marginRight="33dp"
         android:layout_marginBottom="246dp"
+        android:foregroundGravity="center"
         app:srcCompat="@drawable/untitled" />
 
     <TextView
@@ -59,7 +53,6 @@ tools:context=".presenter.game.SecondActivity">
     <Button
         android:id="@+id/startNewGameButton"
         android:layout_width="wrap_content"
-        android:backgroundTint="@color/grey_700"
         android:layout_height="wrap_content"
         android:layout_above="@+id/bottom_navigation"
         android:layout_alignParentStart="true"
@@ -72,6 +65,7 @@ tools:context=".presenter.game.SecondActivity">
         android:layout_marginEnd="118dp"
         android:layout_marginRight="118dp"
         android:layout_marginBottom="35dp"
+        android:backgroundTint="@color/grey_700"
         android:text="@string/start_game_btn"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
@@ -79,14 +73,14 @@ tools:context=".presenter.game.SecondActivity">
         app:layout_constraintTop_toBottomOf="@+id/titleTextView" />
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/bottom_navigation"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:id="@+id/bottom_navigation"
+        android:layout_alignParentBottom="true"
         app:itemBackground="@color/blue_grey_700"
-        app:itemTextColor="@drawable/selector"
         app:itemIconTint="@drawable/selector"
-        app:menu="@menu/menu_navigation"
-        android:layout_alignParentBottom="true"/>
+        app:itemTextColor="@drawable/selector"
+        app:menu="@menu/menu_navigation" />
 
 </RelativeLayout>
