+++ Laboratory1/app/src/main/res/layout/activity_third.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presenter.ThirdActivity">
+    tools:context=".ThirdActivity">
 
     <ImageView
         android:id="@+id/imageView"
